pyproject.toml
[project]
name = "avspl1t"
version = "0.1.0"
description = ""
authors = [
    {name = "Catherine Yeh",email = "catherineyeh@g.harvard.edu"},
    {name = "June Moretz",email = "kmoretz@college.harvard.edu"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pytest (>=8.3.4,<9.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "pytest-dependency (>=0.6.0,<0.7.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "grpcio (>=1.71.0,<2.0.0)",
    "grpcio-tools (>=1.71.0,<2.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "protobuf (>=6.30.2,<7.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false